refactor: move InputText to form directory and seperate Label component from InputText
