+++ packages/form/input-text/public/favicon.ico
