+++ packages/form/input-text/public/manifest.json
