+++ packages/form/input-text/src/react-component-node-modules/components/InputText/InputText.md
@@ -1,0 +1,13 @@
+## titleMaxWidth
+
+```js
+initialState = { value: '' }
+;<div>
+  <InputText
+    type="text"
+    titleMaxWidth={200}
+    value={state.value}
+    onChange={e => setState({ value: e.target.value })}
+  />
+</div>
+```
