+++ packages/form/input-text/src/react-component-node-modules/components/InputText/colors.js
