+++ packages/form/input-text/src/registerServiceWorker.js
