--- packages/input-text/src/react-component-node-modules/components/InputText/InputText.js
@@ -1,111 +1,0 @@
-import React from 'react'
-import styled from 'styled-components'
-import PropTypes from 'prop-types'
-import { BLACK, DARKGRAY, LIGHTGRAY } from './colors'
-
-const Input = styled.input.attrs({
-  type: 'text',
-})`
-  outline: none;
-  border: 1px solid ${DARKGRAY};
-  padding: 0px;
-  min-height: 24px;
-  padding: 0px 0px 0px 10px;
-  &:focus {
-    font-size: 12px;
-    color: ${BLACK};
-    border: 1px solid ${BLACK};
-    caret-color: ${BLACK};
-
-    &::placeholder {
-      color: transparent;
-      background: transparent;
-    }
-  }
-
-  &:disabled {
-    border: 1px solid ${LIGHTGRAY};
-    color: ${LIGHTGRAY};
-    cursor: not-allowed;
-  }
-
-  &::placeholder {
-    padding: 0px;
-    margin: 0px;
-    color: #aeaeae;
-  }
-`
-
-const Label = styled.label`
-  margin-left: ${props => props.titleSpanL}px;
-  margin-right: ${props => props.titleSpanR}px;
-  &:after {
-    content: ' :';
-  }
-`
-
-const Title = styled.div`
-  display: inline-block;
-  max-width: ${props => props.titleMaxWidth}px;
-`
-
-const InputText = ({
-  className,
-  onChange,
-  children,
-  placeholder,
-  value,
-  disabled,
-  titleSpanR,
-  titleSpanL,
-  titleMaxWidth,
-  ...restProps
-}) => {
-  return (
-    <div>
-      <Title titleMaxWidth={titleMaxWidth}>
-        <Label
-          htmlFor={children}
-          titleSpanL={titleSpanL}
-          titleSpanR={titleSpanR}
-        >
-          {children}
-        </Label>
-      </Title>
-      <Input
-        id={children}
-        placeholder={placeholder}
-        value={value}
-        onChange={onChange}
-        className={className}
-        disabled={disabled}
-        {...restProps}
-      />
-    </div>
-  )
-}
-
-InputText.propTypes = {
-  className: PropTypes.string,
-  onChange: PropTypes.func,
-  /** Title name */
-  children: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-  value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-  placeholder: PropTypes.string,
-  disabled: PropTypes.bool,
-  /** The right-width of the title */
-  titleSpanR: PropTypes.number,
-  /** The left-width of the title */
-  titleSpanL: PropTypes.number,
-  /** The max width of the title */
-  titleMaxWidth: PropTypes.number,
-}
-
-InputText.defaultProps = {
-  value: '',
-  disabled: false,
-  titleSpanR: 10,
-  titleSpanL: 0,
-}
-
-export default InputText
